services/availability-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
redis==5.0.1
pydantic==2.6.1
aio-pika==9.4.1
python-dateutil==2.9.0.post0